refactor: moved to constant
